simple/requirements.txt
-e git+git@github.com:carschar/gevent-socketio.git#egg=gevent-socketio
django-extensions

